Move libevent-specific reactor to libevent engine
